+++ test/integration/test_brute_force.py
@@ -358,3 +358,23 @@ class TestBruteForce(TestCaseIntegration):
         expected = self.data_dir/'fuzz/34/brute_force.txt'
         reference = self.data_dir/'downsampled_reference/ENST00000265138.4-ENST00000650150.1'
         self.default_test_case(gvf, reference, expected)
+
+    def test_brute_force_fuzz_35(self):
+        """ Fuzz test 35 """
+        gvf = [
+            self.data_dir/'fuzz/35/fake_variants.gvf',
+            self.data_dir/'fuzz/35/fake_circ_rna.gvf'
+        ]
+        expected = self.data_dir/'fuzz/35/brute_force.txt'
+        reference = self.data_dir/'fuzz/35'
+        self.default_test_case(gvf, reference, expected)
+
+    def test_brute_force_fuzz_36(self):
+        """ Fuzz test 36 """
+        gvf = [
+            self.data_dir/'fuzz/36/fake_variants.gvf',
+            self.data_dir/'fuzz/36/fake_circ_rna.gvf'
+        ]
+        expected = self.data_dir/'fuzz/36/brute_force.txt'
+        reference = self.data_dir/'fuzz/36'
+        self.default_test_case(gvf, reference, expected)